# specs/001-ghostarr-newsletter/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ghostarr API
  description: Newsletter generator for media servers
  version: 1.0.0
  contact:
    name: Ghostarr
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /history:
    get:
      tags: [history]
      summary: List history entries
      operationId: listHistory
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: type
          in: query
          schema:
            type: string
            enum: [manual, automatic]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, success, failed, cancelled]
        - name: template_id
          in: query
          schema:
            type: string
            format: uuid
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Paginated history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryPaginated'
  /history/{id}:
    get:
      tags: [history]
      summary: Get history entry with progress details
      operationId: getHistory
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: History details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'

    delete:
      tags: [history]
      summary: Delete history entry
      operationId: deleteHistory
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
  /history/{id}/regenerate:
    post:
      tags: [history]
      summary: Regenerate with same config
      operationId: regenerateFromHistory
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '202':
          description: Regeneration started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
...
  /history/export:
    get:
      tags: [history]
      summary: Export history
      operationId: exportHistory
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [json, csv]
        - name: type
          in: query
          schema:
            type: string
            enum: [manual, automatic]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, success, failed, cancelled]
      responses:
        '200':
          description: Export file
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoryResponse'
            text/csv:
              schema:
                type: string

  # ============ SETTINGS ============
...
  /logs:
    get:
      tags: [logs]
      summary: List logs
      operationId: listLogs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: level
          in: query
          schema:
            type: string
            enum: [debug, info, warning, error]
        - name: source
          in: query
          schema:
            type: string
            enum: [backend, frontend, integration]
        - name: service
          in: query
          schema:
            type: string
        - name: correlation_id
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsPaginated'

    delete:
      tags: [logs]
      summary: Purge old logs
      operationId: purgeLogs
      parameters:
        - name: older_than_days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Purge result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted_count:
                    type: integer
  /logs/export:
    get:
      tags: [logs]
      summary: Export logs
      operationId: exportLogs
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [json, csv]
        - name: level
          in: query
          schema:
            type: string
            enum: [debug, info, warning, error]
      responses:
        '200':
          description: Export file
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LogResponse'

  # ============ INTEGRATIONS ============
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    GenerationConfig:
      type: object
      required: [template_id]
      properties:
        template_id:
          type: string
          format: uuid
        title:
          type: string
          default: "Newsletter {{date.week}}"
        publication_mode:
          type: string
          enum: [draft, publish, email, email+publish]
          default: draft
        ghost_newsletter_id:
          type: string
          nullable: true
        tautulli:
          $ref: '#/components/schemas/TautulliConfig'
        romm:
          $ref: '#/components/schemas/ContentSourceConfig'
        komga:
          $ref: '#/components/schemas/ContentSourceConfig'
        audiobookshelf:
          $ref: '#/components/schemas/ContentSourceConfig'
        tunarr:
          $ref: '#/components/schemas/TunarrConfig'
        statistics:
          $ref: '#/components/schemas/StatisticsConfig'
        maintenance:
          $ref: '#/components/schemas/MaintenanceConfig'
        max_total_items:
          type: integer
          default: -1
        skip_if_empty:
          type: boolean
          default: false
    ContentSourceConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        days:
          type: integer
          minimum: 1
          maximum: 90
          default: 7
        max_items:
          type: integer
          minimum: -1
          default: -1
    TautulliConfig:
      allOf:
        - $ref: '#/components/schemas/ContentSourceConfig'
        - type: object
          properties:
            featured_item:
              type: boolean
              default: false
    TunarrConfig:
      allOf:
        - $ref: '#/components/schemas/ContentSourceConfig'
        - type: object
          properties:
            days:
              type: integer
              minimum: 1
              maximum: 7
              default: 7
            channels:
              type: array
              items:
                type: string
            display_format:
              type: string
              enum: [grid, list]
              default: grid
    StatisticsConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        days:
          type: integer
          minimum: 1
          maximum: 90
          default: 7
        include_comparison:
          type: boolean
          default: false
    MaintenanceConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        description:
          type: string
        type:
          type: string
          enum: [scheduled, outage, network, update, improvement, security]
          default: scheduled
        duration_value:
          type: integer
          default: 1
        duration_unit:
          type: string
          enum: [hours, days, weeks]
          default: hours
        start_datetime:
          type: string
          format: date-time
          nullable: true

    # ---- Templates ----
    TemplateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        file_path:
          type: string
        preset_config:
          $ref: '#/components/schemas/GenerationConfig'
        is_default:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ScheduleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        is_active:
          type: boolean
        cron_expression:
          type: string
        cron_human_readable:
          type: string
        timezone:
          type: string
        template_id:
          type: string
          format: uuid
        template:
          $ref: '#/components/schemas/TemplateResponse'
        generation_config:
          $ref: '#/components/schemas/GenerationConfig'
        last_run_at:
          type: string
          format: date-time
          nullable: true
        last_run_status:
          type: string
          enum: [pending, success, failed, skipped]
          nullable: true
        next_run_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    HistoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [manual, automatic]
        schedule_id:
          type: string
          format: uuid
          nullable: true
        schedule:
          $ref: '#/components/schemas/ScheduleResponse'
        template_id:
          type: string
          format: uuid
        template:
          $ref: '#/components/schemas/TemplateResponse'
        status:
          type: string
          enum: [pending, running, success, failed, cancelled]
        ghost_post_id:
          type: string
          nullable: true
        ghost_post_url:
          type: string
          nullable: true
        generation_config:
          $ref: '#/components/schemas/GenerationConfig'
        progress_log:
          type: array
          items:
            $ref: '#/components/schemas/ProgressStep'
        error_message:
          type: string
          nullable: true
        items_count:
          type: integer
        duration_seconds:
          type: number
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    HistoryPaginated:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/HistoryResponse'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
    ProgressStep:
      type: object
      properties:
        step:
          type: string
        status:
          type: string
          enum: [pending, running, success, failed, skipped]
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
        items_count:
          type: integer
        message:
          type: string
        error:
          type: string

    # ---- Settings ----
...
    LogResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        level:
          type: string
          enum: [debug, info, warning, error]
        source:
          type: string
          enum: [backend, frontend, integration]
        service:
          type: string
          nullable: true
        message:
          type: string
        context:
          type: object
          nullable: true
        correlation_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    LogsPaginated:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LogResponse'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # ---- Help ----
...